backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pandas==2.1.3
python-multipart==0.0.6
numpy==1.24.3
pymongo==4.6.1

